reboot: not correctly checking host reachability.
Come across the issue where after that the reboot has completed host was
not yet reachable. This is because the host was still booting hence
refusing SSH connection even if SSH service was up and running. wait_for
is the old way to check host reachability and it produced unreliable
results, because the system needs to be booted and we need to make sure
we can execute command over it before continuing the next task.

wait_for_connection does that pretty automatically, it wait for the host
to be reachable and ensure we can run command on it.

Signed-off-by: Lionel H <lionel.hubaut@tessares.net>